services/central-agent/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
redis[hiredis]==5.2.1
asyncpg==0.30.0
neo4j==5.27.0
qdrant-client==1.13.3
prometheus-client==0.21.1
grpcio==1.70.0
grpcio-tools==1.70.0
protobuf==5.29.3
pydantic==2.10.5
httpx==0.28.1
python-dotenv==1.0.1
structlog==24.4.0
psutil==6.1.1
